+++ barony-mod-manager/src/steam_api.rs
@@ -1,92 +1,28 @@
-// use iced::futures::executor::block_on;
-
-// use std::future::Future;
-
-use std::time::Duration;
-
 use iced::{button, image::Handle};
 use reqwest::{Client, Error};
-use serde_json::json;
 
 use crate::{
-    data::{
-        BaronyMod, DownloadStatus, SteamApiResponse, SteamWorkshopMod, SteamWorkshopModResponse,
-        SteamWorkshopTotal,
-    },
+    data::{BaronyMod, DownloadStatus, SteamWorkshopMod},
     filesystem::is_mod_downloaded,
     images::{resize, to_handle},
 };
 
-static BARONY_APP_ID: u64 = 371970;
 static APP_IMAGES_SIZE: u32 = 180; // Pixels
 static DEFAULT_IMAGE: &'static [u8; 4921] = include_bytes!("../resources/img/no_image.png");
 
-pub async fn get_total_mods(client: Client, steam_key: String) -> Result<u64, String> {
-    let params = json!({
-        "key": steam_key,
-        "appid": BARONY_APP_ID.to_string()
-    });
-
-    let response = client
-        .get("https://api.steampowered.com/IPublishedFileService/QueryFiles/v1")
-        .query(&params)
-        .send()
-        .await;
-
-    match response {
-        Ok(response) => {
-            if response.status() != 403 {
-                let json = response
-                    .json::<SteamApiResponse<SteamWorkshopTotal>>()
-                    .await
-                    .unwrap();
-
-                Ok(json.response.total)
-            } else {
-                Err("Not authenticated. Invalid Steam API key.".to_string())
-            }
-        }
-        Err(_message) => {
-            Err("Could not connect to Steam. Is your internet connection up?".to_string())
-        }
-    }
-}
-
-pub async fn get_workshop_item(
+pub async fn get_barony_workshop_mods(
     client: Client,
-    steam_key: String,
-    mod_number: u64,
-) -> Result<SteamWorkshopMod, String> {
-    let params = json!({
-        "key": steam_key,
-        "appid": BARONY_APP_ID,
-        "return_tags": true,
-        "return_vote_data": true,
-        "strip_description_bbcode": true,
-        "page": mod_number
-    });
-
-    // Some sleep so that steam doesn't rate limit us for slapping 100+ concurrent
-    // requests on them.
-    let duration = Duration::from_millis(mod_number * 30);
-    async_std::task::sleep(duration).await;
-
+) -> Result<Vec<SteamWorkshopMod>, reqwest::Error> {
+    let endpoint =
+        "https://raw.githubusercontent.com/arcticlimer/barony-mod-manager/master/data/mods.json";
     let response = client
-        .get("https://api.steampowered.com/IPublishedFileService/QueryFiles/v1")
-        .query(&params)
+        .get(endpoint)
         .send()
-        .await;
+        .await?
+        .json::<Vec<SteamWorkshopMod>>()
+        .await?;
 
-    let mut mod_ = match response {
-        Ok(response) => response
-            .json::<SteamApiResponse<SteamWorkshopModResponse>>()
-            .await
-            .unwrap(),
-        Err(_error) => return Err("Failed to get mod data.".to_string()),
-    };
-
-    let workshop_mod = mod_.response.mods.pop().unwrap();
-    Ok(workshop_mod)
+    Ok(response)
 }
 
 pub async fn build_barony_mod(
@@ -108,11 +44,15 @@ pub async fn build_barony_mod(
     };
 
     BaronyMod {
-        is_downloaded: is_mod_downloaded(&barony_dir, &workshop_data.title.clone()),
-        workshop: workshop_data,
+        // is_downloaded: is_mod_downloaded(&barony_dir, &workshop_data.title.clone()),
         image_handle,
         download_button: button::State::new(),
-        download_status: DownloadStatus::NotDownloaded,
+        download_status: if is_mod_downloaded(&barony_dir, &workshop_data.title.clone()) {
+            DownloadStatus::Downloaded
+        } else {
+            DownloadStatus::NotDownloaded
+        },
+        workshop: workshop_data,
     }
 }